🍀 refactor: 리뷰사항 반영
