+++ app/_types/index.ts
@@ -83,56 +83,3 @@ export type ArtistType = {
   group?: string[];
   profileImage: string;
 };
-
-type ArtistAndGroupType = {
-  id: string;
-  name: string;
-  image: string;
-  type: "solo" | "member" | "group";
-};
-
-export type ArtistAndGroupListType = {
-  page: number;
-  size: number;
-  artistAndGroupList: ArtistAndGroupType[];
-};
-
-type EventImageType = {
-  id: string;
-  eventId: string;
-  imageUrl: string;
-  isMain: boolean;
-};
-
-type TargetArtistsType = {
-  eventId: string;
-  artistId: string;
-  artistName: string;
-  groupId: string;
-  groupName: string;
-};
-
-type eventTagsType = {
-  eventId: string;
-  tagId: string;
-  tagName: GiftType;
-};
-
-export type EventCardType = {
-  id: string;
-  sequence: number;
-  placeName: string;
-  description: string;
-  eventType: EventType;
-  startDate: string;
-  endDate: string;
-  eventUrl: string;
-  organizerSns: string;
-  snsType: string;
-  address: string;
-  addressDetail: string;
-  likeCount: number;
-  eventImages: EventImageType[];
-  targetArtists: TargetArtistsType[];
-  eventTags: eventTagsType[];
-};
